# web/src/api/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "lit-storm",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/discussions": {
   "post": {
    "tags": [
     "discussions"
    ],
    "summary": "Start",
    "operationId": "start_api_discussions_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DiscussionIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DiscussionOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/discussions/{session_id}": {
   "get": {
    "tags": [
     "discussions"
    ],
    "summary": "Get",
    "operationId": "get_api_discussions__session_id__get",
    "parameters": [
     {
      "name": "session_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Session Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DiscussionOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/discussions/{session_id}/turns": {
   "post": {
    "tags": [
     "discussions"
    ],
    "summary": "Ask",
    "operationId": "ask_api_discussions__session_id__turns_post",
    "parameters": [
     {
      "name": "session_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Session Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TurnIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AllowanceOut": {
    "properties": {
     "per_block": {
      "type": "integer",
      "title": "Per Block"
     },
     "blocks": {
      "type": "integer",
      "title": "Blocks"
     },
     "used": {
      "type": "integer",
      "title": "Used"
     },
     "remaining": {
      "type": "integer",
      "title": "Remaining"
     }
    },
    "type": "object",
    "required": [
     "per_block",
     "blocks",
     "used",
     "remaining"
    ],
    "title": "AllowanceOut"
   },
...
   "DiscussionIn": {
    "properties": {
     "topic": {
      "type": "string",
      "maxLength": 500,
      "minLength": 3,
      "title": "Topic"
     },
     "language": {
      "type": "string",
      "pattern": "^(th|en)$",
      "title": "Language"
     },
     "llm_model_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Llm Model Id"
     },
     "search_provider_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Search Provider Id"
     },
     "depth": {
      "type": "string",
      "enum": [
       "fast",
       "standard",
       "deep"
      ],
      "title": "Depth",
      "default": "standard"
     },
     "project_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Project Id"
     },
     "request_key": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 64
       },
       {
        "type": "null"
       }
      ],
      "title": "Request Key"
     }
    },
    "type": "object",
    "required": [
     "topic",
     "language"
    ],
    "title": "DiscussionIn"
   },
   "DiscussionOut": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "project_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Project Id"
     },
     "project_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Project Name"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "language": {
      "type": "string",
      "title": "Language"
     },
     "depth": {
      "type": "string",
      "title": "Depth"
     },
     "model_label": {
      "type": "string",
      "title": "Model Label"
     },
     "search_label": {
      "type": "string",
      "title": "Search Label"
     },
     "turns": {
      "items": {
       "$ref": "#/components/schemas/RunOut"
      },
      "type": "array",
      "title": "Turns"
     },
     "allowance": {
      "$ref": "#/components/schemas/AllowanceOut"
     },
     "current": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/RunOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "view": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "View"
     },
     "reports": {
      "items": {
       "$ref": "#/components/schemas/ReportRef"
      },
      "type": "array",
      "title": "Reports"
     }
    },
    "type": "object",
    "required": [
     "id",
     "project_id",
     "project_name",
     "title",
     "language",
     "depth",
     "model_label",
     "search_label",
     "turns",
     "allowance",
     "current",
     "view",
     "reports"
    ],
    "title": "DiscussionOut"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "QuestionAnswer": {
    "properties": {
     "question": {
      "type": "string",
      "maxLength": 300,
      "title": "Question"
     },
     "answer": {
      "type": "string",
      "maxLength": 1000,
      "title": "Answer",
      "default": ""
     }
    },
    "type": "object",
    "required": [
     "question"
    ],
    "title": "QuestionAnswer"
   },
...
   "ReportRef": {
    "properties": {
     "run_id": {
      "type": "string",
      "title": "Run Id"
     },
     "title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     },
     "source_count": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Count"
     },
     "finished_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Finished At"
     }
    },
    "type": "object",
    "required": [
     "run_id",
     "title",
     "source_count",
     "finished_at"
    ],
    "title": "ReportRef"
   },
...
   "RunOut": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "session_id": {
      "type": "string",
      "title": "Session Id"
     },
     "parent_run_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Parent Run Id"
     },
     "topic": {
      "type": "string",
      "title": "Topic"
     },
     "language": {
      "type": "string",
      "title": "Language"
     },
     "engine": {
      "type": "string",
      "title": "Engine"
     },
     "engine_label": {
      "type": "string",
      "title": "Engine Label"
     },
     "model_label": {
      "type": "string",
      "title": "Model Label"
     },
     "search_label": {
      "type": "string",
      "title": "Search Label"
     },
     "depth": {
      "type": "string",
      "title": "Depth"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "quota_month": {
      "type": "string",
      "title": "Quota Month"
     },
     "stage": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Stage"
     },
     "reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reason"
     },
     "message": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Message"
     },
     "quota_refunded": {
      "type": "boolean",
      "title": "Quota Refunded"
     },
     "queued_at": {
      "type": "string",
      "format": "date-time",
      "title": "Queued At"
     },
     "started_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Started At"
     },
     "finished_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Finished At"
     },
     "report_title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Report Title"
     },
     "source_count": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Count"
     },
     "notes": {
      "additionalProperties": true,
      "type": "object",
      "title": "Notes",
      "default": {}
     },
     "refinement": {
      "items": {
       "$ref": "#/components/schemas/QuestionAnswer"
      },
      "type": "array",
      "title": "Refinement",
      "default": []
     },
     "turn": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Turn"
     },
     "quota_units": {
      "type": "integer",
      "title": "Quota Units",
      "default": 1
     }
    },
    "type": "object",
    "required": [
     "id",
     "session_id",
     "parent_run_id",
     "topic",
     "language",
     "engine",
     "engine_label",
     "model_label",
     "search_label",
     "depth",
     "status",
     "quota_month",
     "stage",
     "reason",
     "message",
     "quota_refunded",
     "queued_at",
     "started_at",
     "finished_at",
     "report_title",
     "source_count"
    ],
    "title": "RunOut"
   },
...
   "TurnIn": {
    "properties": {
     "action": {
      "type": "string",
      "enum": [
       "say",
       "step",
       "auto",
       "report",
       "start"
      ],
      "title": "Action"
     },
     "text": {
      "type": "string",
      "maxLength": 2000,
      "title": "Text",
      "default": ""
     },
     "steps": {
      "type": "integer",
      "maximum": 10.0,
      "minimum": 1.0,
      "title": "Steps",
      "default": 3
     },
     "extend": {
      "type": "boolean",
      "title": "Extend",
      "default": false
     },
     "request_key": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 64
       },
       {
        "type": "null"
       }
      ],
      "title": "Request Key"
     }
    },
    "type": "object",
    "required": [
     "action"
    ],
    "title": "TurnIn"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}